Struct nanomsg::Endpoint

pub struct Endpoint<'a> {
    // some fields omitted
}

An endpoint created for a specific socket. Each endpoint is identified by a unique return value that can be further passed to a shutdown function. The shutdown is done through the endpoint itself and not the Socket. However, the Endpoint doesn't live longer than the socket itself. This is done through phantom lifetimes.

Methods

impl<'a> Endpoint<'a>

fn new(value: c_int, socket: c_int) -> Endpoint<'a>

fn shutdown(&'a mut self) -> NanoResult<()>