Struct nanomsg::PollFd

pub struct PollFd {
    // some fields omitted
}

A request for polling a socket and the poll result. To create the request, see Socket::new_pollfd. To get the result, see PollFd::can_read and PollFd::can_write.

Methods

impl PollFd

fn can_read(&self) -> bool

Checks whether at least one message can be received from the socket without blocking.

fn can_write(&self) -> bool

Checks whether at least one message can be sent to the fd socket without blocking.