Struct nanomsg::PollRequest

pub struct PollRequest<'a> {
    // some fields omitted
}

A request for polling a set of sockets and the poll results. To create the request, see PollRequest::new.

Methods

impl<'a> PollRequest<'a>

fn new(fds: &'a mut [PollFd]) -> PollRequest<'a>

Creates a request from the specified individualsocket requests.

fn get_fds(&'a self) -> &'a [PollFd]

Returs a reference to the socket requests, so they can be checked.