Crate nanomsg

Reexports

pub use result::{Result, Error};

pub use endpoint::Endpoint;

Modules

endpoint

result

Structs

PollFd
A request for polling a socket and the poll result. To create the request, see Socket::new_pollfd. To get the result, see PollFd::can_read and PollFd::can_write.

PollRequest
A request for polling a set of sockets and the poll results. To create the request, see PollRequest::new.

Socket
A type-safe socket wrapper around nanomsg's own socket implementation. This provides a safe interface for dealing with initializing the sockets, sending and receiving messages.

Enums

PollInOut

Protocol
Type-safe protocols that Nanomsg uses. Each socket is bound to a single protocol that has specific behaviour (such as only being able to receive messages and not send them).

Transport