Struct nanomsg::endpoint::Endpoint

pub struct Endpoint {
    // some fields omitted
}

An endpoint created for a specific socket. Each endpoint is identified by a unique return value that can be further passed to a shutdown function. The shutdown is done through the endpoint itself and not the Socket

Methods

impl Endpoint

fn new(value: c_int, socket: c_int) -> Endpoint

fn shutdown(&mut self) -> Result<()>

Removes an endpoint from the socket that created it (via bind or connect). The call will return immediately, however, the library will try to deliver any outstanding outbound messages to the endpoint for the time specified by Socket::set_linger.