Enum nanomsg::Protocol

pub enum Protocol {
    Req,
    Rep,
    Push,
    Pull,
    Pair,
    Bus,
    Pub,
    Sub,
    Surveyor,
    Respondent,
}

Type-safe protocols that Nanomsg uses. Each socket is bound to a single protocol that has specific behaviour (such as only being able to receive messages and not send 'em).

Variants

Req

Rep

Push

Pull

Pair

Bus

Pub

Sub

Surveyor

Respondent