Struct nanomsg::Socket

pub struct Socket<'a> {
    // some fields omitted
}

A type-safe socket wrapper around nanomsg's own socket implementation. This provides a safe interface for dealing with initializing the sockets, sending and receiving messages.

Methods

impl<'a> Socket<'a>

fn new(protocol: Protocol) -> NanoResult<Socket<'a>>

Allocate and initialize a new Nanomsg socket which returns a new file descriptor behind the scene. The safe interface doesn't expose any of the underlying file descriptors and such.

Usage:

use nanomsg::{Socket, Protocol};

let mut socket = match Socket::new(Protocol::Pull) {
    Ok(socket) => socket,
    Err(err) => panic!("{}", err)
};

fn new_for_device(protocol: Protocol) -> NanoResult<Socket<'a>>

fn bind<'b, 'a>(&mut self, addr: &str) -> NanoResult<Endpoint<'b>>

Creating a new socket through Socket::new does not bind that socket to a listening state. Instead, one has to be explicit in enabling the socket to listen onto a specific address.

That's what the bind method does. Passing in a raw string like: "ipc:///tmp/pipeline.ipc" is supported.

Note: This does not block the current task. That job is up to the user of the library by entering a loop.

Usage:

use nanomsg::{Socket, Protocol};

let mut socket = match Socket::new(Protocol::Pull) {
    Ok(socket) => socket,
    Err(err) => panic!("{}", err)
};

// Bind the newly created socket to the following address:
//match socket.bind("ipc:///tmp/pipeline.ipc") {
//    Ok(_) => {},
//   Err(err) => panic!("Failed to bind socket: {}", err)
//}

fn connect<'b, 'a>(&mut self, addr: &str) -> NanoResult<Endpoint<'b>>

fn nb_read(&mut self, buf: &mut [u8]) -> NanoResult<uint>

fn nb_read_to_end(&mut self) -> NanoResult<Vec<u8>>

fn nb_write(&mut self, buf: &[u8]) -> NanoResult<()>

fn new_pollfd(&self, pollin: bool, pollout: bool) -> PollFd

fn poll(request: &mut PollRequest, timeout: &Duration) -> NanoResult<int>

fn device(socket1: &Socket, socket2: &Socket) -> NanoResult<()>

fn terminate()

fn set_linger(&mut self, linger: &Duration) -> NanoResult<()>

fn set_send_buffer_size(&mut self, size_in_bytes: int) -> NanoResult<()>

fn set_receive_buffer_size(&mut self, size_in_bytes: int) -> NanoResult<()>

fn set_send_timeout(&mut self, timeout: &Duration) -> NanoResult<()>

fn set_receive_timeout(&mut self, timeout: &Duration) -> NanoResult<()>

fn set_reconnect_interval(&mut self, interval: &Duration) -> NanoResult<()>

fn set_max_reconnect_interval(&mut self, interval: &Duration) -> NanoResult<()>

fn set_send_priority(&mut self, priority: u8) -> NanoResult<()>

fn set_receive_priority(&mut self, priority: u8) -> NanoResult<()>

fn set_ipv4_only(&mut self, ipv4_only: bool) -> NanoResult<()>

fn set_socket_name(&mut self, name: &str) -> NanoResult<()>

fn set_tcp_nodelay(&mut self, tcp_nodelay: bool) -> NanoResult<()>

fn subscribe(&mut self, topic: &str) -> NanoResult<()>

fn unsubscribe(&mut self, topic: &str) -> NanoResult<()>

fn set_survey_deadline(&mut self, deadline: &Duration) -> NanoResult<()>

fn set_request_resend_interval(&mut self, interval: &Duration) -> NanoResult<()>